Refactor Brain class, limit scoping
Make neurons vector private and limit scope of other variables where
possible. This was acheived by removing dim and n_neurons as members
of the class itself and instead passing them to the member functions
when needed, which in the case of n_neurons was only a single time.

dim was never used at all as it can be obtained from the length of
the bounds valarray.