Correct signal propagation
Was missing removal of live node from set, this meant that the set
quickly was filled with all nodes in the network and hence no new
nodes could be added to the queue.

In addition, was missing removal of nodes from the set/queue when
propagation was not successful (eg not enough signal enter a neuron
to cause a firing). This meant that the set/queue grew larger than
they should and this had a significant effect on performance.

Added comments to explain the whole procedure.